+++ buildpacks/add-cnb-require-jdk/src/main.rs
@@ -1,0 +1,32 @@
+// Enable rustc and Clippy lints that are disabled by default.
+// https://doc.rust-lang.org/rustc/lints/listing/allowed-by-default.html#unused-crate-dependencies
+#![warn(unused_crate_dependencies)]
+// https://rust-lang.github.io/rust-clippy/stable/index.html
+#![warn(clippy::pedantic)]
+
+use libcnb::build::{BuildContext, BuildResult, BuildResultBuilder};
+use libcnb::buildpack_main;
+use libcnb::data::build_plan::BuildPlanBuilder;
+use libcnb::detect::{DetectContext, DetectResult, DetectResultBuilder};
+use libcnb::generic::{GenericError, GenericMetadata, GenericPlatform};
+use libcnb::Buildpack;
+
+struct AddCnbRequireJdkBuildpack;
+
+impl Buildpack for AddCnbRequireJdkBuildpack {
+    type Platform = GenericPlatform;
+    type Metadata = GenericMetadata;
+    type Error = GenericError;
+
+    fn detect(&self, _context: DetectContext<Self>) -> libcnb::Result<DetectResult, Self::Error> {
+        DetectResultBuilder::pass()
+            .build_plan(BuildPlanBuilder::new().requires("jdk").build())
+            .build()
+    }
+
+    fn build(&self, _context: BuildContext<Self>) -> libcnb::Result<BuildResult, Self::Error> {
+        BuildResultBuilder::new().build()
+    }
+}
+
+buildpack_main!(AddCnbRequireJdkBuildpack);